test: DCI 패턴을 적용한 AuthServiceTest 로 대체
